# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gestion des Utilisateurs",
    "version": "1.0.0",
    "description": "API pour la gestion des utilisateurs, y compris l'authentification et la manipulation des données utilisateur."
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/promotions": {
      "post": {
        "summary": "Créer une nouvelle promotion",
        "description": "Permet aux utilisateurs de créer une nouvelle promotion.",
        "operationId": "createPromotion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libelle": {
                    "type": "string",
                    "example": "Promotion 2"
                  },
                  "date_debut": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-01"
                  },
                  "duree": {
                    "type": "integer",
                    "example": 6
                  },
                  "referentiels": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "AWS001"
                        },
                        "libelle": {
                          "type": "string",
                          "example": "AWS"
                        },
                        "description": {
                          "type": "string",
                          "example": "claud"
                        },
                        "duree_acquisition": {
                          "type": "string",
                          "example": "3 mois"
                        },
                        "statut": {
                          "type": "string",
                          "example": "ACTIF"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "libelle",
                  "date_debut"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Promotion créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "promotion": {
                      "type": "object",
                      "description": "Détails de la promotion créée"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation des données"
          }
        }
      },
      "get": {
        "summary": "Récupérer toutes les promotions",
        "description": "Permet de récupérer une liste de toutes les promotions existantes.",
        "operationId": "getAllPromotions",
        "responses": {
          "200": {
            "description": "Liste des promotions récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "promotionId123"
                      },
                      "libelle": {
                        "type": "string",
                        "example": "Promotion 6"
                      },
                      "date_debut": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                      },
                      "date_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-01"
                      },
                      "duree": {
                        "type": "integer",
                        "example": 6
                      },
                      "etat": {
                        "type": "string",
                        "example": "Actif"
                      },
                      "referentiels": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "example": "AWS001"
                            },
                            "description": {
                              "type": "string",
                              "example": "claud"
                            },
                            "duree_acquisition": {
                              "type": "string",
                              "example": "3 mois"
                            },
                            "etat": {
                              "type": "string",
                              "example": "Actif"
                            },
                            "libelle": {
                              "type": "string",
                              "example": "AWS"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/promotions/{id}": {
      "patch": {
        "summary": "Mettre à jour une promotion existante",
        "description": "Permet de mettre à jour les détails d'une promotion spécifique.",
        "operationId": "updatePromotion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "L'ID de la promotion à mettre à jour",
            "schema": {
              "type": "string",
              "example": "promotionId123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date_debut": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-01"
                  },
                  "date_fin": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-07-01"
                  },
                  "duree": {
                    "type": "integer",
                    "example": 6
                  },
                  "etat": {
                    "type": "string",
                    "example": "Actif"
                  },
                  "libelle": {
                    "type": "string",
                    "example": "Promotion 6"
                  },
                  "referentiels": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "AWS001"
                        },
                        "description": {
                          "type": "string",
                          "example": "claud"
                        },
                        "duree_acquisition": {
                          "type": "string",
                          "example": "3 mois"
                        },
                        "etat": {
                          "type": "string",
                          "example": "Actif"
                        },
                        "libelle": {
                          "type": "string",
                          "example": "AWS"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "date_debut",
                  "libelle",
                  "referentiels"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promotion mise à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "promotion": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Promotion not found."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation des données"
          }
        }
      }
    },
    "/api/v1/promotions/{id}/referentiels": {
      "patch": {
        "summary": "Mettre à jour les référentiels d'une promotion",
        "description": "Permet de mettre à jour les référentiels associés à une promotion existante.",
        "operationId": "updatePromotionReferentiels",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la promotion à mettre à jour",
            "schema": {
              "type": "string",
              "example": "promotionId123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date_debut": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-01"
                  },
                  "date_fin": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-07-01"
                  },
                  "duree": {
                    "type": "integer",
                    "example": 6
                  },
                  "etat": {
                    "type": "string",
                    "example": "Actif"
                  },
                  "libelle": {
                    "type": "string",
                    "example": "Promotion 6"
                  },
                  "referentiels": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "AWS008"
                        },
                        "description": {
                          "type": "string",
                          "example": "claud et DevOps"
                        },
                        "duree_acquisition": {
                          "type": "string",
                          "example": "5 mois"
                        },
                        "etat": {
                          "type": "string",
                          "example": "Actif"
                        },
                        "libelle": {
                          "type": "string",
                          "example": "AWS"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Référentiels mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Référentiels mis à jour."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Promotion not found."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Récupérer les référentiels pour une promotion donnée",
        "description": "Renvoie la liste des référentiels actifs associés à la promotion spécifiée par son ID.",
        "operationId": "getPromotionReferentiels",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "L'ID de la promotion pour laquelle récupérer les référentiels.",
            "schema": {
              "type": "string",
              "example": "promo123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Référentiels récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string",
                        "example": "AWS001"
                      },
                      "libelle": {
                        "type": "string",
                        "example": "AWS"
                      },
                      "description": {
                        "type": "string",
                        "example": "claud et DevOps"
                      },
                      "duree_acquisition": {
                        "type": "string",
                        "example": "5 mois"
                      },
                      "etat": {
                        "type": "string",
                        "example": "ACTIF"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Aucune promotion trouvée pour l'ID spécifié",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Promotion not found."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/promotions/{id}/etat": {
      "patch": {
        "summary": "Mettre à jour l'état d'une promotion",
        "description": "Permet de mettre à jour l'état d'une promotion existante.",
        "operationId": "updatePromotionEtat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la promotion dont l'état doit être mis à jour",
            "schema": {
              "type": "string",
              "example": "promotionId123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date_debut": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-01-01"
                  },
                  "date_fin": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-07-01"
                  },
                  "duree": {
                    "type": "integer",
                    "example": 6
                  },
                  "etat": {
                    "type": "string",
                    "example": "Actif"
                  },
                  "libelle": {
                    "type": "string",
                    "example": "Promotion 6"
                  },
                  "referentiels": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "AWS001"
                        },
                        "description": {
                          "type": "string",
                          "example": "claud et DevOps"
                        },
                        "duree_acquisition": {
                          "type": "string",
                          "example": "5 mois"
                        },
                        "etat": {
                          "type": "string",
                          "example": "Actif"
                        },
                        "libelle": {
                          "type": "string",
                          "example": "AWS"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "État de la promotion mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "État de la promotion mis à jour."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Promotion not found."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/promotions/encours": {
      "get": {
        "summary": "Récupérer la promotion active",
        "description": "Renvoie la promotion qui est actuellement active.",
        "operationId": "getActivePromotion",
        "responses": {
          "200": {
            "description": "Promotion active récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "PROMO001"
                    },
                    "libelle": {
                      "type": "string",
                      "example": "Promotion 6"
                    },
                    "date_debut": {
                      "type": "string",
                      "format": "date",
                      "example": "2024-01-01"
                    },
                    "date_fin": {
                      "type": "string",
                      "format": "date",
                      "example": "2024-07-01"
                    },
                    "duree": {
                      "type": "integer",
                      "example": 6
                    },
                    "etat": {
                      "type": "string",
                      "example": "Actif"
                    },
                    "referentiels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string",
                            "example": "AWS001"
                          },
                          "libelle": {
                            "type": "string",
                            "example": "AWS"
                          },
                          "description": {
                            "type": "string",
                            "example": "claud et DevOps"
                          },
                          "duree_acquisition": {
                            "type": "string",
                            "example": "5 mois"
                          },
                          "etat": {
                            "type": "string",
                            "example": "Actif"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Aucune promotion active trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No active promotion found."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}